Spark 4.1 - handle partial row updates defaulting missing columns
