+++ spark-3.1-spanner-lib/src/test/java/com/google/cloud/spark/spanner/integration/SchemaValidationIntegrationTestBase.java
@@ -124,6 +124,26 @@ public class SchemaValidationIntegrationTestBase extends SparkSpannerIntegration
   }
 
   @Test
+  public void testFullWriteSucceedsWithoutOption() {
+    StructType partialSchema =
+        new StructType(
+            new StructField[] {
+              DataTypes.createStructField("id", DataTypes.LongType, false),
+              DataTypes.createStructField("name", DataTypes.StringType, true),
+              DataTypes.createStructField("value", DataTypes.DoubleType, true),
+            });
+    List<Row> rows = Collections.singletonList(RowFactory.create(1L, "test", 1.23));
+    Dataset<Row> df = spark.createDataFrame(rows, partialSchema);
+
+    System.out.println("df.queryExecution().analyzed() " + df.queryExecution().analyzed());
+
+    Map<String, String> props = connectionProperties(usePostgreSql);
+    props.put("table", SCHEMA_VALIDATION_TABLE_NAME);
+    // "enablePartialRowUpdates" is NOT set
+    df.write().format("cloud-spanner").options(props).mode(SaveMode.Append).save();
+  }
+
+  @Test
   public void testNumericScaleMismatchFailsInGoogleSql() {
     if (usePostgreSql) {
       // This validation is specific to GoogleSQL's strict NUMERIC(38,9) type.
