+++ spark-4.1-spanner-lib/src/test/java/com/google/cloud/spark/spanner/integration/Spark41SchemaValidationIntegrationTest.java
@@ -56,24 +56,4 @@ public class Spark41SchemaValidationIntegrationTest extends SchemaValidationInte
     String message = e.getMessage();
     assertThat(message).contains("Partial row updates require enablePartialRowUpdates=true.");
   }
-
-  @Test
-  public void testFullWriteSucceedsWithoutOption() {
-    StructType partialSchema =
-        new StructType(
-            new StructField[] {
-              DataTypes.createStructField("id", DataTypes.LongType, false),
-              DataTypes.createStructField("name", DataTypes.StringType, true),
-              DataTypes.createStructField("value", DataTypes.DoubleType, true),
-            });
-    List<Row> rows = Collections.singletonList(RowFactory.create(1L, "test", 1.23));
-    Dataset<Row> df = spark.createDataFrame(rows, partialSchema);
-
-    System.out.println("df.queryExecution().analyzed() " + df.queryExecution().analyzed());
-
-    Map<String, String> props = connectionProperties(usePostgreSql);
-    props.put("table", SCHEMA_VALIDATION_TABLE_NAME);
-    // "enablePartialRowUpdates" is NOT set
-    df.write().format("cloud-spanner").options(props).mode(SaveMode.Append).save();
-  }
 }
